% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{chrom_ranges_from_lengths}
\alias{chrom_ranges_from_lengths}
\title{Calculate chromosome ranges from chromosome lengths}
\usage{
chrom_ranges_from_lengths(chr_lengths)
}
\description{
Calculate chromosome ranges from chromosome lengths
}
