% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{coverage_depth_summary}
\alias{coverage_depth_summary}
\title{Counts the number of bases covered by 0, 1, 2 etc CNV intervals}
\usage{
coverage_depth_summary(data, chrlengths, endpos_is_inclusive)
}
\arguments{
\item{data}{- data.table of CNV intervals}

\item{chrlengths}{- data.table of chromosome lengths}

\item{endpos_is_inclusive}{- TRUE means an interval [start=10, end=20] includes the last position,
so is 11 bases long (i.e. 10, 11, 12, ..., 19, 20]. FALSE means an interval [start=10, end=20]
excludes the last position, and is 10 bases long, because it stops at 19.}
}
\description{
Counts the number of bases covered by 0, 1, 2 etc CNV intervals
}
