% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{depth_of_coverage}
\alias{depth_of_coverage}
\title{Calculates the depth of coverage (of CNV segments) for the genome,
discretised into bins of size `binsize`}
\usage{
depth_of_coverage(data, chrom_lengths, binsize = 1e+05)
}
\description{
Calculates the depth of coverage (of CNV segments) for the genome,
discretised into bins of size `binsize`
}
