% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{do_simulation}
\alias{do_simulation}
\title{Simulation code}
\usage{
do_simulation(input_data, chrlengths, simulation_name, nsims, base_outdir,
  write_output = FALSE, keep_to_original_chrom = FALSE,
  hard_bounds = FALSE, parametric = FALSE)
}
\arguments{
\item{input_data}{data.table; table of CNVs}

\item{chrlengths}{data.table; table of the Devil chromosome lengths}

\item{simulation_name}{string; used in name of output files}

\item{nsims}{int; number of simulation replicates}

\item{base_outdir}{path; output will be written under this path}

\item{write_output}{bool; write output to disk?}

\item{keep_to_original_chrom}{bool; should shuffled CNV segments be restricted
to their original chromosome, or be allowed to move genome wide?}

\item{hard_bounds}{bool; should chromosome ends act as hard boundaries for shuffled
CNVs? Or should CNVs be allowed to straddle over chromosome ends?}

\item{parametric}{bool; if TRUE, then CNV widths are drawn from a Gamma distribution,
fitted to the observed widths in the input data. If FALSE, the original widths
are used.}
}
\description{
Simulation code
}
