% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_aberrant_copy_number_per_sample}
\alias{get_aberrant_copy_number_per_sample}
\title{Make a table of aberrant copy number states for each combination of CNV ID & sample ID.
'Aberrant' copy number is returned as NA if the CNV is not present in the sample, which
doesn't mean that the copy number is 0, it means that the copy number is not aberrant,
and is the inherited wild-type state of the DFTD founder Devil (most likely 2).}
\usage{
get_aberrant_copy_number_per_sample(cnv_table)
}
\arguments{
\item{cnv_table}{data.table; Table of CNV information}
}
\description{
Make a table of aberrant copy number states for each combination of CNV ID & sample ID.
'Aberrant' copy number is returned as NA if the CNV is not present in the sample, which
doesn't mean that the copy number is 0, it means that the copy number is not aberrant,
and is the inherited wild-type state of the DFTD founder Devil (most likely 2).
}
