% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_depth}
\alias{get_depth}
\title{Calculate number of bases covered by a table of CNVs
Start and end coordinates must be in the format that the end coordinate
is the position of the first base NOT covered by the CNV (SAM format)}
\usage{
get_depth(data_table, chrom_lengths = NULL)
}
\description{
Calculate number of bases covered by a table of CNVs
Start and end coordinates must be in the format that the end coordinate
is the position of the first base NOT covered by the CNV (SAM format)
}
