% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_variant_site}
\alias{is_variant_site}
\title{Check if a vector of characters is a variant site}
\usage{
is_variant_site(chars, alphabet = c("DNA", "BINARY"))
}
\description{
Check if a vector of characters is a variant site
}
