% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_categorisation.R
\name{isMultiGain}
\alias{isMultiGain}
\title{Return a vector of CNV IDs associated with a copy number gain of >1}
\usage{
isMultiGain(cnvtable)
}
\description{
Return a vector of CNV IDs associated with a copy number gain of >1
}
