% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_categorisation.R
\name{isMultiLoss}
\alias{isMultiLoss}
\title{Return a vector of CNV IDs associated with a copy number loss of < -1}
\usage{
isMultiLoss(cnvtable)
}
\description{
Return a vector of CNV IDs associated with a copy number loss of < -1
}
