% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_cnv_table}
\alias{load_cnv_table}
\title{Loads the CNV table (supplementary table S3) in data.table format.}
\usage{
load_cnv_table()
}
\description{
Loads the CNV table (supplementary table S3) in data.table format.
}
