% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_sample_table}
\alias{load_sample_table}
\title{Loads the tumour sample table (supplementary table S1A) in data.table format.}
\usage{
load_sample_table()
}
\description{
Loads the tumour sample table (supplementary table S1A) in data.table format.
}
