% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_coordinates}
\alias{make_coordinates}
\title{Helper for making matching segment tables when looking at recurrent breakpoints}
\usage{
make_coordinates(cnv_table)
}
\arguments{
\item{cnv_table}{data.table; Table of CNV information}
}
\description{
Helper for making matching segment tables when looking at recurrent breakpoints
}
