% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_presence_absence_table}
\alias{make_presence_absence_table}
\title{Make a presence / absence table for CNVs
Melts the CNV table on sample name. The output table has the columns
"New.CNV_ID_ext", "chr", "start", "end", "Index", "Type", plus any
extra columns passed in the `extra_columns` parameter}
\usage{
make_presence_absence_table(cnv_table, extra_columns = NULL)
}
\arguments{
\item{cnv_table}{data.table; Table of CNV information}

\item{extra_columns}{character vector; vector of additional columns to include in output.}
}
\description{
Make a presence / absence table for CNVs
Melts the CNV table on sample name. The output table has the columns
"New.CNV_ID_ext", "chr", "start", "end", "Index", "Type", plus any
extra columns passed in the `extra_columns` parameter
}
