% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pad}
\alias{pad}
\title{Pad vector v to length padlen, using padval as the added value
If v is already as long as or longer than padlen, return v
(this function does not truncate)}
\usage{
pad(v, padlen, padval = 0)
}
\description{
Pad vector v to length padlen, using padval as the added value
If v is already as long as or longer than padlen, return v
(this function does not truncate)
}
