% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_control.R
\name{parse_sample_cnv_group_and_count}
\alias{parse_sample_cnv_group_and_count}
\title{parse sample cnv group and count}
\usage{
parse_sample_cnv_group_and_count(s)
}
\description{
parse sample cnv group and count
}
