% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{preprocess_for_simulation}
\alias{preprocess_for_simulation}
\title{Data cleaning ready for making simulations}
\usage{
preprocess_for_simulation(cnvtable, chrom_lengths_file)
}
\description{
Data cleaning ready for making simulations
}
