% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{randomize_positions_over_whole_genome}
\alias{randomize_positions_over_whole_genome}
\title{Takes a data frame of segments and randomises their start
positions (up to max start), while keeping their width constant.
If the end position exceeds maxstart, the segment is 'wrapped around'
by cutting into two segments, the second of which is mapped to position 0.}
\usage{
randomize_positions_over_whole_genome(segments, limit)
}
\description{
Takes a data frame of segments and randomises their start
positions (up to max start), while keeping their width constant.
If the end position exceeds maxstart, the segment is 'wrapped around'
by cutting into two segments, the second of which is mapped to position 0.
}
