% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangle_packing.R
\name{rectangle_packing2}
\alias{rectangle_packing2}
\title{Algorithm to pack segments into layers for plotting. Greedily assigns widest segments first.}
\usage{
rectangle_packing2(dt)
}
\description{
Algorithm to pack segments into layers for plotting. Greedily assigns widest segments first.
}
