% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectangle_packing.R
\name{rectangle_packing3}
\alias{rectangle_packing3}
\title{A faster algorithm to pack aligned reads into layers}
\usage{
rectangle_packing3(dt, sortby = "width")
}
\description{
A faster algorithm to pack aligned reads into layers
}
