% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome_map_plotting_functions.R,
%   R/rectangle_packing.R
\name{rectangle_packing4}
\alias{rectangle_packing4}
\title{A faster algorithm to pack aligned reads into layers. Doesn't need IRanges.}
\usage{
rectangle_packing4(dt, sortby = "width")

rectangle_packing4(dt, sortby = "width")
}
\description{
A faster algorithm to pack aligned reads into layers. Doesn't need IRanges.

A faster algorithm to pack aligned reads into layers. Doesn't need IRanges.
}
