% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_control.R
\name{scan_col}
\alias{scan_col}
\title{Easier extraction of data from data table columns using regex}
\usage{
scan_col(pattern, s, dtypes = NULL, names = NULL)
}
\description{
Easier extraction of data from data table columns using regex
}
