% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{wrap_around}
\alias{wrap_around}
\title{Wrap segments that exceed the limit value back into the start of the sequence}
\usage{
wrap_around(dt, limit)
}
\description{
Wrap segments that exceed the limit value back into the start of the sequence
}
