% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_fasta}
\alias{write_fasta}
\title{Write sequence information in a table to FASTA format}
\usage{
write_fasta(filename, table)
}
\description{
Write sequence information in a table to FASTA format
}
