% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome_map_plotting_functions.R
\name{plot_chromosome_map}
\alias{plot_chromosome_map}
\title{Plot an individual chromosome's map. Assumes dt.gains and dt.losses
have already been produced, including setting the plotting layers
for each segment with rectangle packing}
\usage{
plot_chromosome_map(dt_gains, dt_losses, chr_lengths,
  plot_options = NULL, get_plot_options = FALSE, chr_override = NULL)
}
\description{
Plot an individual chromosome's map. Assumes dt.gains and dt.losses
have already been produced, including setting the plotting layers
for each segment with rectangle packing
}
